+++ presentation/src/main/java/com/threegap/bitnagil/presentation/home/HomeScreen.kt
@@ -28,8 +28,8 @@ import com.threegap.bitnagil.presentation.home.component.template.CollapsibleHom
 import com.threegap.bitnagil.presentation.home.component.template.EmptyRoutineView
 import com.threegap.bitnagil.presentation.home.component.template.RoutineSection
 import com.threegap.bitnagil.presentation.home.component.template.WeeklyDatePicker
-import com.threegap.bitnagil.presentation.home.model.HomeSideEffect
-import com.threegap.bitnagil.presentation.home.model.HomeState
+import com.threegap.bitnagil.presentation.home.contract.HomeSideEffect
+import com.threegap.bitnagil.presentation.home.contract.HomeState
 import com.threegap.bitnagil.presentation.home.util.rememberCollapsibleHeaderState
 import org.orbitmvi.orbit.compose.collectAsState
 import org.orbitmvi.orbit.compose.collectSideEffect
